removing examples from footer & header + implementation of sidenav in app layout
