+++ website/screens/components/footer/usage/FooterUsagePage.tsx
@@ -6,16 +6,12 @@ import {
 import DocFooter from "@/common/DocFooter";
 import QuickNavContainer from "@/common/QuickNavContainer";
 import QuickNavContainerLayout from "@/common/QuickNavContainerLayout";
-import Example from "@/common/example/Example";
-import usage from "./examples/usage";
-import content from "./examples/content";
 
 const sections = [
   {
     title: "Usage",
     content: (
       <>
-        <Example example={usage} />
         <DxcBulletedList>
           <DxcBulletedList.Item>
             The footer frame should be docked at the bottom of the page and
@@ -44,7 +40,6 @@ const sections = [
     title: "Content",
     content: (
       <>
-        <Example example={content} />
         <DxcParagraph>
           The footer component has a custom area where many kinds of content can
           be placed, some of them are contemplated in the following list:
