--- website/screens/components/header/code/examples/dropdown.tsx
@@ -1,43 +1,0 @@
-import { DxcHeader, DxcInset, DxcButton } from "@dxc-technology/halstack-react";
-
-const code = `() => {
-  const selectOption = (value) => {
-    console.log(value);
-  };
-  const options = [
-    {
-      value: 1,
-      label: "Amazon",
-    },
-    {
-      value: 2,
-      label: "Ebay",
-    },
-    {
-      value: 3,
-      label: "Apple",
-    },
-  ];
-
-  return (
-    <DxcInset space="2rem">
-      <DxcHeader
-        content={
-          <DxcHeader.Dropdown
-            options={options}
-            onSelectOption={selectOption}
-            label="Store"
-          />
-        }
-      />
-    </DxcInset>
-  );
-}`;
-
-const scope = {
-  DxcHeader,
-  DxcInset,
-  DxcButton,
-};
-
-export default { code, scope };
